Improve deployment performance on large clusters
This change makes the nova key distribution a lot faster
especially when deploying against very large clusters. The
change moves away from the authorized_key module and instead
generates a script with the same capabilities which is then
executed. The generated script is stored in locally on the remote
host at "/usr/local/bin/openstack-nova-key.sh" and can be
executed at any time to fix and or clean up nova authorized key
problems.

Change-Id: I0d5ec9d735a104a57ec5cf7938116915af803088
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>