Fix shell command usage
In https://review.openstack.org/320624 a method of providing the
shell module arguments was used which breaks ansible-lint (and
is improper anyway). The lint check did not effectively catch
this. The lint check not working will be resolved in another
patch. This patch fixes the task's syntax to ensure that fixed
lint checks implemented in other repositories do not fail on
this role.

Change-Id: If62fed46937193c0868ca2c9d2b4c29bfd331c59