Only install to virtual environment
Remove all tasks and variables related to toggling between installation
of nova inside or outside of a Python virtual environment.
Installing within a venv is now the only supported deployment.

Additionally, a few changes have been made to make the creation of the
venv more resistant to interruptions during a run of the role.
* unarchiving a pre-built venv will now also occur when the venv
  directory is created, not only after being downloaded
* virtualenv-tools is run against both pre-built and non pre-built venvs
  to account for interruptions during or prior to unarchiving

Change-Id: Ia8687d49803705702f0e9f7ea9a83d3c05694d46
Implements: blueprint only-install-venvs