Move UCA enable and repo URL var to role defaults
In order to expose the var in role documentation and to allow the use
of dynamically set facts to override the value in CI environments the
variables are moved from the role vars to the role defaults.

The test preparation implements an override of the URL for OpenStack-CI
to make use of the local mirrors.

Change-Id: Iffa32adb971f281e7e274209ee500b1756fbf0a1