If /var/lib/libvirt/qemu/save exists, move it to nova_libvirt_save_path
In https://review.openstack.org/320624 the qemu save directory was set
to be a link into the same directory where nova instances are kept in
order to ensure that temporary files created in snapshots are stored
in a location which is likely on shared storage or has plenty of space.

The patch did not actually create the link if a folder existed there
already, so this patch implements a move of the folder and a link to
replace it.

Closes-Bug: #1585325
Change-Id: I5224ac72b14ad027410f5d5c14d20870dd37b5cb