Fix novnc console include positioning
Currently the nova console (spice/novnc) includes happen before the
venv validation/download in nova_install. While this works for spice
(the default) because it doesn't use the nova venv, novnc does, which
breaks nova console install when novnc is enabled. This patch moves
the include for novnc (and spice) to after the venv checksum/download.

Change-Id: I8c6fb525727c177ef69d8ac5fa915840643a12a8