Fix race condition for nova_compute installation
The nova gate is failing intermittently because the libvirt-python pip
package is not installed on localhost (compute host) consistently.

This seems to be caused by the dyanmic inclusion on the compute
installation tasks. The pip installation task is skipped altogether on
the failed gates, whilst it is run on the successful gates.

Change-Id: I482e6a66999da8194eafbd57eae1b1f422c8e193