Ensure that novnc/spice consoles use the public endpoint protocol
When using HTTPS as the public endpoint protocol, the novnc/spice
console endpoint provided must also be HTTPS.

This patch ensures that the novnc/spice console endpoint keys off
the general OpenStack service protocol set for the public endpoints.

It still remains possible to override the endpoint protocol, if
necessary.

Closes-Bug: #1630950
Closes-Bug: #1630953
Change-Id: If3c751adfc4cb74c3230db1c8d4f1c9c3672bea8