Delegate from first available compute host
The 'Drop authorized keys file script locally' task shouldn't
specifically require that the first member of the nova_compute group be
available. If the playbook is run with a limit excluding that host, the
task will always end up being skipped.

Use an intersect filter to find the first available compute host within
the play and use that as the delegation host.

Closes-Bug: 1632447
Change-Id: Id12fd6cee104a6fb705acc66704f21d629694d84