Nova upgrade testing.
Implement upgrade testing for os_nova role.

This patch adds an upgrade testenv, which configures a previous version
of Nova, and the test-vars from the previous stable branch to be used
by the previous version of Nova.

We don't run functional tests after the initial stable/newton deploy
because a working stable/newton branch is already tested, and it's
assumed it worked if the upgrade works too - which is more
efficient.

This follows the pattern setup for Keystone in:
https://review.openstack.org/#/c/384269/

TODO: (andymccr)
Nova deprecated the nova_cert project, which means the group has been
removed from the os_nova role. For now a workaround has been added to
add the nova_cert group to the upgrades inventory, the default is to
have nova_cert in it's own container so we could just remove those as
part of an N->O Upgrade.

Change-Id: I3a83c3f91c85ea532d1963c007bb81cae6de2cd7
Implements: blueprint upgrade-testing