Enable cell auto enrollment
When discover_hosts_in_cells_interval is disabled, we need to wait
for the nova-compute service to register itself in the service
database or else it may not be started in time for cell enrollment
to occur.

However, when automatic enrollment is enabled in the scheduler, we
do not need to wait for the service to start because the scheduler
will enroll it automatically using a periodic task.

By default we use an automatic enrollment interval of 60 seconds when 10
or less compute hosts or present, or 300 seconds when >10 compute hosts
are in the deployment.

Change-Id: I02bb69c5f39b4686402bb815bdd0b978bcb05b06