Setup "os_interface" option for placement service
Since https://review.openstack.org/#/c/426163/ has now merged we can
move to use the os_interface option in the [placement] section of
nova.conf.

Setting this to "internal" will ensure we utilize the internal keystone
endpoint instead of the defaulted public endpoint.

We can now remove the work-around to set the public endpoint to use the
internal_lb_vip address and protocol.

Change-Id: Ib8753f99a1c445207bfff6059d76eba150d69b5b