Reload nginx instead of restart
We don't need to restart nginx - we can instead just reload the
service. Additionally, as more services move to use nginx frontend, it
would be bad to restart all NGinx services at the same time.

For now we need to investigate the impact of reloads in UWsgi before
moving over to a "reload" on UWsgi.

Change-Id: I60e370e784a1ff3a0f5bf8551be804bf05d8bb43