Remove trailing slash in nginx config
The nginx configuration file passes along traffic to the placement
API if the URI is '/placement/' but the service is registered with
keystone as just '/placement' (no trailing slash). This causes
nginx to send back the default page when nova services try to talk
to the placement API.

This patch fixes the nginx configuration to match the URI for the
registered service.

Closes-Bug: 1689824
Change-Id: I309730a6daa8c3ea15fca78ab9ce910383019869