Remove warning about duplicate item var
Because the task file is in a loop include with item name, we
should ensure we're not using "item" in another place.
This is renaming the lower level task instead of the upper level
task (we should probably do that in the future, as it is a good
practice), to fix the current problem in the libguestfs linking:

[WARNING]: The loop variable 'item' is already in use. You should set
the `loop_var` value in the `loop_control` option for the task to
something else to avoid variable collisions and unexpected behavior.

Change-Id: I9eb0f12a3693b8c923557c519cef429a747e498b