Normalise distro package installation
This patch makes the distro package list compilation
just a little easier to understand, and avoids the
double evaluation of the variable. It also adds the
spice distro package list which was missing.

It also passes the package list directly to the package
module instead of using a with_items loop and removes
the nested inclusions for distro package installation.

Finally, the async process for installing the packages
is removed as it doesn't give much benefit in terms of
speed and instead obscures the process a little.

Change-Id: I910868db9780d9b56eabf5f059fe2f6d32719238