Use direct include instead of "with_first_found"
We specify the full path and are not searching for a generic file, we
should rather use a direct include instead of using "with_first_found".

Additionally, we need the include paths within drivers/virt_type to be
relative rather than absolute - otherwise these files fail to be
included properly.

Change-Id: Ie9bd6d778d134bcba28f19a5e7895ccb8bb26da9