Correct nova virt type detection
The nova role was attempting to detect the virt type if one was not
already pre-set however the tasks would only ever set the virt type to
"qemu" which forces all workloads to run in software instead of
hardware virt which has massive impacts in terms of workload
performance. This change corrects the virt type detection logic so that
it will run fall back to qemu should another virt type not be detected.

Change-Id: I78093808f03238c3c019797a06bba3ceca821239
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>