Avoid explicit requirement for groups to exist
The current templating for nova.conf checks if the 'inventory_hostname'
of the current host is found within the combined lists of hosts for the
default groups of all services other than nova-compute. This requires
that each of those explicitly mentioned groups exist in the inventory
and future upkeep if any of those default groups are renamed, added to,
or removed.

Instead, lookup each group defined in the 'nova_services' variable for
all services other than nova-compute and check for an intersection of
those groups and the current host's 'group_names'.

Depends-On: https://review.openstack.org/540613
Change-Id: Ib3fe93e4f6a9f473a599cd4ac3e96ee9ac919108