Rearrange cell mapping tasks
Existing instances can be mapped to a newly created cell immediately
after the cell is created and by the same conductor host that created
the cell.

After a new compute host is built, it will need to mapped to a cell
through the 'discover_hosts' command of nova-manage. This command will
still need to be run against a host that has direct database access, but
the command is able to handle mapping of all compute hosts that do not
currently belong to the cell so it only needs to be run once per play.

There can be a short delay in the time between a compute service
starting and becoming available to be discovered and mapped but, for
most deployments, compute host discovery will be handled automatically
as a periodic task.

Change-Id: I4f7328e1aff79eb94f278999b97fbad8f44c6a3e
Closes-Bug: 1749037
Related-Bug: 1729661
Related-Bug: 1752540