Restore openstack-ansible-upgrade-ubuntu-xenial
Without this patch, nova upgrades are not tested.
As a reminder, Master and Queens were simulatenously broken
together due to the SPICE url change. We can now bring the
upgrades back now that the SPICE url change has been merged
in Queens [1].

The fact we do not test upgrades is a problem, as we could
miss an important upstream change, or break our upgrades
ourselves without noticing.

This patches solves the problem.

[1]: I5639ebee2d232fdac3f5fe2d38e9a54d2ff8c03a

Change-Id: I31e874ecad4525e57b85f2b09a556124a8f8829d