Drop usage of nova_spice_distro_packages in the role
There was a commit 3 years ago which switched to installing SPICE
from source rather than from packages.  It also included a section
which would 'remove' and 'cleanup' packages for SPICE which was
not removed.

With time, it looks like we've added it to the list of things to
install, only to uninstall it in the next task.  This removes all
of this extra cruft and further removes our dependency on EPEL.

Change-Id: I9bbe109d50040bbc02b194f9486ab581a6cc51a9