Setup service users and endpoints before database setup
As part of the nova-status upgrade check process, it makes sure that
the credentials setup for the service are working.  Because we run
that in nova_db_setup.yml, it means that the endpoints and users
don't exist yet and it cannot validate it.

This instead sets up the service much earlier in order to make sure
that Nova can find it.

Change-Id: Idf14240a9843a55575fd1816c78ba9d4a8a19632