Fix ordering of source/console installs
When the console is set to 'novnc', the install process
will conditionally install pip packages into the venv
when in developer mode. To detect this is checks for
'nova_get_venv | failed or nova_get_venv | skipped' which
is only present if the source-install has been done first.

This patch fixes the ordering so that the installation of
the venv is done before the console install.

Change-Id: I3fc4bba4398d5aec605cb3ba19ae482c72f94bca