Ensure python3 compatibility
Currently everything is wired in the venv to work on python2.
If the virtualenv was built during repo build with python3, the
linking of the library will fail, as the lib folder for python2
doesn't exist.

This should solve the issue by ensuring the existing lib folder
of the env is used.

A problem arises then: Different distros organise packages
differently, in different locations. This makes the approach
compatible with multi-distros and multiple python versions.

Change-Id: I02b296089d5351d68bb14e87e458a4f5d18a4776