Retry nova-status check and only run it once
In https://review.openstack.org/575125 the nova-statuc check was
implemented to run on the control plane nodes to verify a healthy
environment.

Instead of running on all control plane nodes, we set it to only
run on the first one to ensure that the play stops if it fails
which will hopefully leave a running environment, giving time for
deployers to fix it.

Also, sometimes the API is not yet ready, so we implement some
retries to ensure that extra time is given for any API's that need
it.

Finally, the task is renamed to make it more descriptive to those
who are unfamiliar with what nova-status is and its purpose.

Change-Id: Iea8a71653df5a8506b0e29f587585d9dfb5a7a1b