If nova is installed on arm64 with active KVM, default to that.
There are problems using aarch64 booting UEFI with bare QEMU at the
moment, but if run under KVM it works well with a 4.15.0 kernel.

Tested with the current Ubuntu Bionic cloud image.  The Cirros 0.4.0
arm64 image has a 4.4.0 kernel which doesn't receive hardware
information from UEFI and panics when it cannot find an interrupt
controller, but without logging anything to console because it hasn't
detected the serial port either.

Change-Id: I7a39a03bf1bf959c650e4454bcb32f8c0c792619