Always deploy placement authtoken config
The placement service is required, nova cannot start without it,
so we must always configure the authtoken even if the flag to deploy
the integrated nova-placement-api is disabled. This will enable the
ability to move placement into a separate service, disable the
nova-placement-api deployment, yet still configure the required
authtoken in nova.conf.

Depends-On: I58788af6f02a4b339a270dff69a86ce7cdea41d3
Change-Id: I4619d333187a0e96c1b60fe2f203df0838c12059