tasks: add conditionals for ironic
This patch avoids running the nova_compute.yml playbook
entirely because we already skip the setup portion and we
don't need to setup SSH keys (as there's no live migration)
nor setting up NFS.

It also gets rid of the set_fact inside nova_virt_detect.yml
because it was non-functional in the first place.  This is
because the nova_virt_type was set to 'ironic' in the
group_vars which meant that the autodetect code was
skipped and therefore that portion never ran.

Change-Id: Ic5548f08d200402538a0334f801834ee7ef8a39b