Disable unused console services
This is best practise to only deploy one remote access method at a
time. This patch ensures only one of the nova-novncproxy, nova-spicehtml5proxy
or nova-serialproxy services is enabled and running.

This is a necessary step for upgrades from stein to train which use
the OSA defaults, where the console type has will switch from spice to novnc.
The old nova-spicehtml5proxy service continually fail and restart, running
from the stein venv.

The service fails because the nova config option use_journal is now
set to true. Nova-spicehtml5proxy service consumes this config and tries to
use journal logging which is not possible in the stein venv due to a 
missing systemd-python package. This dependancy was only added in Train.

Change-Id: I420a95caf7981034f6549c2b6f8296839d9ed65f