Make possible to set nova_libvirt_images_rbd_pool to empty string
Checking if nova_libvirt_images_rbd_pool is defined leaves us without
an opportunity to disable ceph usage for specific node.
Instead of that we will use bool filter, which will be true if line
is not empty, and false if it is. That way by setting variable to empty
string in host_vars we can disable all relevant options for specific node

Change-Id: Id7c03399bd54ac790dcfc28ad1dcd7f6d48979db