Add Virtual GPU Config to nova.conf template
Currently, the list of supported vGPUs is populated automatically
within the Nova playbook when the /sys/class/mdev_bus folder is
present. This will populate the `enabled_vgpu_types` variable with
the full list of vGPUs that are supported by default, with only
the first one listed being used.

This change provides the ability to define a new variable
`nova_enabled_vgpu_types` that will populate the enabled_vgpu_types
config option. It also adds the ability to map vgpu types to PCI
addresses in line with Ussuri updates to provide multi-vGPU support.

Change-Id: Icc0c2cd896d4c9a01601d4d733f38443ad8400d4