Set default qemu settings for RBD
When Nova and/or Cinder are using Ceph as backend, qemu will need
to open a connection and two threads for each and every Ceph OSD.
Since all connections occur in the same qemu process, this may result
in hitting default max open files limit.
Thus in case of more then 10 volumes are attached
to the same instance and are used actively can end up in blocked
IO operations inside VMs.

We increase these limits by default when RBD backend is used.

Change-Id: Ib3081280cdbae1eb2235083c95c27e2efd0b413e