Use template URLs in Cell Mappings
In Rocky Nova has implemented templating in Cell Mappings [1]
That means that instead of hardcoding connection details in database,
we can just put a template there. Variables are parsed from nova.conf
and substituted on the service load. Thus we don't need to update cells
every time we change passoword and we can use different credentials
across nodes since they will be just taken from config files.

We also perform upgrade of the cells to the templates when cell exist.

Task `Map instances to new Cell1` is removed, since it's required to
map instances to cellv2 only during upgrade from cellsv1 to cellsv2
which is not the case nowadays or when migrating instances between
cells [2].

[1] https://docs.openstack.org/nova/latest/user/cells.html#template-urls-in-cell-mappings
[2] https://docs.openstack.org/nova/rocky/cli/nova-manage.html#nova-cells-v2
Change-Id: Ia6bef7b902c0fb99a529c592172226bb16ed0d9d