Allow to provide mdev addresses as list
With new Ampere GPUs you need to explicitly define explicit lists of PCI
devices which length depends on picked type as placement can't pick them
automatically due to nvidia driver brokeness.

In order to have readable representation of the variable it's worth
to make it iterable but keep a simple string for
backwards compatability.

Change-Id: I2a1e85efc8ad4f6a2596e6d53b1d793b2f934758