Stop installing qemu-system on debian variants
qemu-system on debian derivative OS is a meta-package which installs
qemu-system-* for all architecures understood by qemu.

This is different from redhat type OS where the qemu-kvm package
installed with dnf only installs the qemu-system-* binary matching
the host architecture.

This gives two problems, first there is inconsistency in openstack-ansible
deployments between redhat and debian OS. Second, there is a potentially
unexpected emulation of architectures when launching VM on a cloud
with a mix of compute architectures when a full set of qemu-system-*
binaries is available on a compute node. The compute node becomes a
candidate for scheduling any of the supported architectures and a
very specific configuration is needed both from the operator and end
user to ensure that VM are run on a native architecture or emulated as
required.

This patch changes the installation so that redhat and debian compute nodes
only have the native qemu-system binary installed.

A new feature should be introduced to openstack-ansible in the future
to explicitly control installation of non-native qemu-system-* binaries
and write the config options for controlling emulation.

Change-Id: I1c876c7968efb7f24880f1a6e96ba6b7264ddc94